backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
boto3>=1.35.75
requests-oauthlib>=2.0.0
cryptography>=43.0.3
python-dotenv>=1.0.1
pymongo==4.10.1
pydantic>=2.10.3
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.6.0
pytest>=8.3.4
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.13.0
python-jose>=3.3.0
requests>=2.32.3
aiosmtplib>=3.0.0
pandas>=2.2.3
numpy>=2.0.2
python-multipart>=0.0.20
jq>=1.8.0
typer>=0.15.1
